Phase 5 commit 4: extract notes/hooks/{useCaretTracking,useLinkPicker}.ts
- useLinkPicker owns the @-mention dropdown state machine: linkPicker,
  linkPickerRef, maybeShowLinkPicker (caret → maybe-show), filteredLinks
  memo, pickLink (insert + caret restore via rAF). 124 lines extracted.
- useCaretTracking owns activeLineRef and refreshActive, plus the
  document selectionchange listener that re-renders the active-line
  marker and pings the picker via onCaretMove callback.

LiveEditor.tsx 434 → 301 lines. Build clean, tests baseline unchanged.
